Home Page front end beginning
I have created the shell for the Homepage.

If you want to navigate to it, the url is /homepage at the end of the local host url. You will need to run server and all of that to test it.

Eric and Nic, the changes I have made will allow you to add and change HTML and CSS style code in those file locations. This is what you need to do to make the front end more appealing - while I work through and start devving the process for the other user webpages amongst other things.